Factor out password logic from the general form
The password logic and UI (that is, the options and length change logic) has
been moved out of the main password generator form, into a separate
function.

This makes the generator more easily composable later when a passphrase
(which also needs its own options) is added to the form.
This way, the UI can swap out the options section when either password
or passphrase is selected.